combined some detectors to custom_type.ts
