+++ packages/netsuite-adapter/src/changes_detector/changes_detector.ts
@@ -18,12 +18,10 @@ import _ from 'lodash'
 import { SuiteAppClient } from '../client/suiteapp_client/suiteapp_client'
 import { NetsuiteQuery } from '../query'
 import { getChangedFiles, getChangedFolders } from './changes_detectors/file_cabinet'
-import customRecordTypeDetector from './changes_detectors/custom_record_type'
-import customFieldDetector from './changes_detectors/custom_field'
+import { customRecordTypeDetector, customFieldDetector, customListDetector } from './changes_detectors/custom_type'
 import scriptDetector from './changes_detectors/script'
 import roleDetector from './changes_detectors/role'
 import workflowDetector from './changes_detectors/workflow'
-import customListDetector from './changes_detectors/custom_list'
 import savedSearchDetector from './changes_detectors/savedsearch'
 import { formatSavedSearchDate } from './formats'
 import { ChangedType, DateRange } from './types'
