+++ packages/netsuite-adapter/test/changes_detector/changes_detector.test.ts
@@ -17,7 +17,7 @@
 import { SuiteAppClient } from '../../src/client/suiteapp_client/suiteapp_client'
 import { NetsuiteQuery } from '../../src/query'
 import * as fileCabinetDetector from '../../src/changes_detector/changes_detectors/file_cabinet'
-import customRecordTypeDetector from '../../src/changes_detector/changes_detectors/custom_record_type'
+import { customRecordTypeDetector } from '../../src/changes_detector/changes_detectors/custom_type'
 import scriptDetector from '../../src/changes_detector/changes_detectors/script'
 import { getChangedObjects } from '../../src/changes_detector/changes_detector'
