+++ packages/netsuite-adapter/test/changes_detector/custom_record_type.test.ts
@@ -14,7 +14,7 @@
 * limitations under the License.
 */
 import { SuiteAppClient } from '../../src/client/suiteapp_client/suiteapp_client'
-import detector from '../../src/changes_detector/changes_detectors/custom_record_type'
+import { customRecordTypeDetector as detector } from '../../src/changes_detector/changes_detectors/custom_type'
 import { Change } from '../../src/changes_detector/types'
 
 describe('custom_record_type', () => {
